feat(Kitchenpos): refactor product package
