+++ src/test/java/kitchenpos/menus/bo/MenuBoTest.java
@@ -4,8 +4,8 @@ import kitchenpos.menus.dao.MenuDao;
 import kitchenpos.menus.dao.MenuGroupDao;
 import kitchenpos.menus.dao.MenuProductDao;
 import kitchenpos.menus.model.Menu;
-import kitchenpos.products.bo.InMemoryProductDao;
-import kitchenpos.products.dao.ProductDao;
+import kitchenpos.products.application.InMemoryProductDao;
+import kitchenpos.products.infrastructure.dao.ProductDao;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
